Moving Section1 to a module by itself
